improved the solid layers duplication
as solid source duplication is either unsupported or unknown to me, added a process of creating a solid layer, with exactly the same parameters as the layer to be duplicated, which itself creates a new source: the main reason for creating a solid. Then changing the solid layer's source to that new source of the new layer, then deleting the new layer.